Move Rulesets into GitHub plugin
