+++ default_config.yaml
@@ -1,5 +1,4 @@
-includes: ["GitHub", "GitHubCommunityStandards", "GitHubRulesets"]
+includes: ["GitHub", "GitHubCommunityStandards"]
 GitHub-url: "https://github.com/gt-sse-center/RepoAuditor"
 GitHubCommunityStandards-url: "https://github.com/gt-sse-center/RepoAuditor"
-GitHubRulesets-url: "https://github.com/gt-sse-center/RepoAuditor"
 # GitHub-pat: "PAT"
