+++ pyproject.toml
@@ -41,7 +41,6 @@ repo_auditor = "RepoAuditor:EntryPoint.app"
 [project.entry-points.RepoAuditor]
 GitHubPlugin = "RepoAuditor.Plugins.GitHubPlugin"
 GitHubCommunityStandardsPlugin = "RepoAuditor.Plugins.GitHubCommunityStandardsPlugin"
-GitHubRulesetPlugin="RepoAuditor.Plugins.GitHubRulesetsPlugin"
 
 [build-system]
 requires = ["hatchling"]
