+++ src/RepoAuditor/Plugins/GitHub/RulesetRequirements/RequireSignedCommits.py
@@ -6,7 +6,7 @@
 # -------------------------------------------------------------------------------
 """Contains the RequireSignedCommits object."""
 
-from RepoAuditor.Plugins.GitHubRulesets.Impl.EnableRulesetRequirementImpl import EnableRulesetRequirementImpl
+from RepoAuditor.Plugins.GitHub.Impl.EnableRulesetRequirementImpl import EnableRulesetRequirementImpl
 
 
 class RequireSignedCommits(EnableRulesetRequirementImpl):
